Don't poll if pomo is paused or stopped
The command inotifywait allows to "wait" for changes on a file or
directory. This way pomo stops polling if it is in the paused or
stopped state.

Afaik it is not possible to use inotifywait for files which don't exist.
That means inotifywait has to listen on the folder to detect whether a
pomo went from state stopped to started. The defaul folder is
$HOME/.local/share which is to generic to listen for change events. So
I've added a subfolder pomo and renamed the file from pomo to stat.